More refactor and Media Edit fix relates to xibosignageltd/xibo-private#1230
