+++ frontend/src/config/commonForms.ts
@@ -1,0 +1,40 @@
+/*
+ * Copyright (C) 2026 Xibo Signage Ltd
+ *
+ * Xibo - Digital Signage - https://xibosignage.com
+ *
+ * This file is part of Xibo.
+ *
+ * Xibo is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * any later version.
+ *
+ * Xibo is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with Xibo.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+export const COMMON_FORM_OPTIONS = {
+  orientation: [
+    { label: 'Portrait', value: 'portrait' },
+    { label: 'Landscape', value: 'landscape' },
+    { label: 'Square', value: 'square' },
+  ],
+  inherit: [
+    { label: 'Off', value: 'off' },
+    { label: 'On', value: 'on' },
+    { label: 'Inherit', value: 'inherit' },
+  ],
+  lastModifiedFilter: [
+    { label: 'Any time', value: '' },
+    { label: 'Today', value: 'today' },
+    { label: 'Last 7 days', value: '7d' },
+    { label: 'Last 30 days', value: '30d' },
+    { label: 'This year', value: '1y' },
+  ],
+};
